learning/mongodb/rag/requirements.txt
openai
markdown
pymongo
python-dotenv
trulens_eval
numpy
datasets
llama_index
llama-index-vector-stores-mongodb
IPython